feat: Add a config module
